// backend/src/domain/models/physical-assessment.ts
export interface PhysicalAssessmentModel {
  id: string
  user: string
  weight: number
  height: number
  rightBiceps: number
  leftBiceps: number
  rightForearm: number
  leftForearm: number
  chest: number
  waist: number
  abdomen: number
  rightThigh: number
  leftThigh: number
  rightCalf: number
  leftCalf: number
  date: string
  responsible: string
}